+++ src/Test/MuCheck/TestAdapter.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE DeriveDataTypeable #-}
 -- | Module for adapting test framekworks
 module Test.MuCheck.TestAdapter where
 
@@ -5,18 +6,22 @@ import qualified Language.Haskell.Interpreter as I
 import Data.Typeable
 
 -- | Wrapper for interpreter output
-type InterpreterOutput a = (Either I.InterpreterError (String, a), String)
+data Summarizable a => InterpreterOutput a = Io {_io :: Either I.InterpreterError a, _ioLog::String}
 
 -- | Holding mutant information
 type Mutant = String
 
+-- | Holding test information
+type TestStr = String
+
 -- | Summary of test run
 newtype Summary = Summary String
+  deriving (Show, Typeable)
 
 -- | Interface to be implemented by a test framework
 class Typeable s => Summarizable s where
-  -- | Summary of a test run
-  testSummary :: [Mutant] -> [InterpreterOutput s] -> Summary
+  -- | Summary of test suite on a single mutant
+  testSummary :: Mutant -> TestStr -> InterpreterOutput s -> Summary
   -- | Was the test run a success
   isSuccess :: s -> Bool
   -- | Was the test run a failure