+++ src/Test/MuCheck/Utils/Print.hs
@@ -41,8 +41,7 @@ catchOutputStr f = do
 
 -- | Capture output and err of an IO action to a file
 catchOutput :: String -> IO a -> IO a
-catchOutput fn f = do
-  withFile fn WriteMode (redirectToHandle f)
+catchOutput fn f = withFile fn WriteMode (redirectToHandle f)
 
 -- | Redirect out and err to handle
 redirectToHandle :: IO b -> Handle -> IO b